extract checkDistribution into its own file, simplify it and make it more general
