+++ packages/grader/src/checks/axiom-checks.ts
@@ -227,175 +227,6 @@ export const mulOne: Check = (prev, next, context) => {
 };
 mulOne.symmetric = true;
 
-export const checkDistribution: Check = (prev, next, context) => {
-    // Handle the situation where we have a term within an 'add' node that needs
-    // distributing.
-    if (prev.type === "add" && next.type === "add") {
-        const results: Result[] = [];
-
-        // Find all 'mul' nodes and then try generating a newPrev node from
-        // each of them.
-        for (let i = 0; i < prev.args.length; i++) {
-            const mul = prev.args[i];
-
-            const potentialResults: Result[] = [];
-            const potentialNewPrevs = [];
-
-            if (
-                mul.type === "mul" &&
-                mul.args.length === 2 &&
-                mul.args[1].type === "add"
-            ) {
-                const newPrev = Semantic.builders.add([
-                    ...prev.args.slice(0, i),
-                    ...mul.args[1].args.map((arg) =>
-                        Semantic.builders.mul([mul.args[0], arg], mul.implicit),
-                    ),
-                    ...prev.args.slice(i + 1),
-                ]);
-
-                const result = context.checker.checkStep(
-                    newPrev,
-                    next,
-                    context,
-                );
-                if (result) {
-                    potentialResults.push(result);
-                    potentialNewPrevs.push(newPrev);
-                }
-            } else if (
-                mul.type === "mul" &&
-                mul.args.length === 2 &&
-                mul.args[0].type === "add"
-            ) {
-                const newPrev = Semantic.builders.add([
-                    ...prev.args.slice(0, i),
-                    ...mul.args[0].args.map((arg) =>
-                        Semantic.builders.mul([arg, mul.args[1]], mul.implicit),
-                    ),
-                    ...prev.args.slice(i + 1),
-                ]);
-
-                const result = context.checker.checkStep(
-                    newPrev,
-                    next,
-                    context,
-                );
-                if (result) {
-                    potentialResults.push(result);
-                    potentialNewPrevs.push(newPrev);
-                }
-            }
-
-            if (potentialResults.length > 0) {
-                let result = potentialResults[0];
-                let newPrev = potentialNewPrevs[0];
-                if (potentialResults.length === 2) {
-                    if (
-                        potentialResults[1].steps.length < result.steps.length
-                    ) {
-                        result = potentialResults[1];
-                        newPrev = potentialNewPrevs[1];
-                    }
-                }
-                results.push(
-                    correctResult(
-                        prev,
-                        newPrev,
-                        context.reversed,
-                        [],
-                        result.steps,
-                        "distribution",
-                        "factoring",
-                    ),
-                );
-            }
-        }
-
-        // If there are multiple results, pick the one with the shortest number
-        // of steps.
-        if (results.length > 0) {
-            let shortestResult = results[0];
-            for (const result of results.slice(1)) {
-                if (result.steps.length < shortestResult.steps.length) {
-                    shortestResult = result;
-                }
-            }
-            return shortestResult;
-        }
-    }
-
-    if (prev.type !== "mul" || next.type !== "add") {
-        return;
-    }
-
-    const potentialResults: Result[] = [];
-    const potentialNewPrevs = [];
-
-    // If the second factor is an add, e.g. a(b + c) -> ...
-    if (prev.args[1].type === "add") {
-        const newPrev = Semantic.builders.add(
-            prev.args[1].args.map((arg) => {
-                if (arg.type === "neg") {
-                    // Set 'subtraction' prop to false
-                    return Semantic.builders.mul([
-                        prev.args[0],
-                        Semantic.builders.neg(arg.arg),
-                    ]);
-                } else {
-                    return Semantic.builders.mul(
-                        [prev.args[0], arg],
-                        prev.implicit,
-                    );
-                }
-            }),
-        );
-
-        const result = context.checker.checkStep(newPrev, next, context);
-        if (result) {
-            potentialResults.push(result);
-            potentialNewPrevs.push(newPrev);
-        }
-    }
-
-    // If the first factor is an add, e.g. (b + c)a -> ...
-    if (prev.args[0].type === "add") {
-        const newPrev = Semantic.builders.add(
-            prev.args[0].args.map((arg) =>
-                Semantic.builders.mul([arg, prev.args[1]]),
-            ),
-        );
-
-        const result = context.checker.checkStep(newPrev, next, context);
-        if (result) {
-            potentialResults.push(result);
-            potentialNewPrevs.push(newPrev);
-        }
-    }
-
-    if (potentialResults.length > 0) {
-        let result = potentialResults[0];
-        let newPrev = potentialNewPrevs[0];
-        if (potentialResults.length === 2) {
-            if (potentialResults[1].steps.length < result.steps.length) {
-                result = potentialResults[1];
-                newPrev = potentialNewPrevs[1];
-            }
-        }
-        return correctResult(
-            prev,
-            newPrev,
-            context.reversed,
-            [],
-            result.steps,
-            "distribution",
-            "factoring",
-        );
-    }
-};
-
-checkDistribution.symmetric = true;
-
 export const mulByZero: Check = (prev, next, context) => {
     const {checker} = context;
